refactor(layout-editor): colocate page and component support code
