+++ docs/engineering.md
@@ -43,6 +43,10 @@ Local-first collaboration:
   default style for local store updates.
 - Use simple normalization such as `value ?? null` unless real runtime type
   narrowing is required.
+- In handlers, destructure the dependencies you use at the top of the
+  function (`const { store, refs, appService, projectService } = deps`)
+  instead of repeatedly using `deps.store`, `deps.refs`, and similar dotted
+  access throughout the body.
 
 ## Testing
 
@@ -157,6 +161,12 @@ High-level rules:
 - stores hold UI-local state and derived display data
 - handlers orchestrate
 - page `*.store.js` and `*.handlers.js` remain the composition root for a page
+- page-private non-visual support code should stay in the owning page folder
+  under `support/`
+- page-local visual workflows with their own surface and state should usually
+  move into dedicated components under `src/components/`
+- components must not import from other component folders; shared code belongs
+  in page `support/` or another non-component shared layer
 - `src/internal/ui/` owns shared page/store/handler orchestration
 - `src/internal/project/` owns project meaning and invariants
 - `src/internal/project/` stays intentionally merged into a small number of
@@ -166,42 +176,6 @@ High-level rules:
 - `src/deps/services/` owns service behavior and service adapters
 - `src/deps/clients/` owns low-level platform/external clients
 - setup entry points create dependencies
-
-### UI Feature Slice Pattern
-
-For long-lived page workflows, prefer feature slices over growing one large
-page store or handler file.
-
-Good examples:
-
-- canvas interaction
-- preview-state orchestration
-- dialog/create flows
-
-Recommended shape:
-
-```text
-src/internal/ui/<page-or-domain>/<feature>.js
-```
-
-That feature module may own, together in one file:
-
-- initial state
-- store actions
-- selectors
-- handlers
-- RxJS stream/subscription helpers
-
-Use this pattern when those pieces are tightly coupled and splitting them into
-feature-local `*.store.js` and `*.handlers.js` files would only add
-navigation overhead.
-
-Keep these boundaries:
-
-- page `*.store.js` and `*.handlers.js` stay the composition root
-- cross-feature orchestration stays at page level
-- pure project meaning still belongs in `src/internal/project/`
-- service behavior still belongs behind `appService` / `projectService`
 
 ## Repository Schema Alignment
 
@@ -327,6 +301,11 @@ scene-specific workflows.
 
 - `src/pages/`
   Route-level screens and screen-specific orchestration.
+  Page-private supporting code that is not a component and not shared across
+  pages should live in `src/pages/<page-name>/support/`.
+  Keep `support/` flat. Do not create nested subfolders under `support/`.
+  If that code becomes a real UI workflow or visual surface, promote it into a
+  component instead of adding more support-folder hierarchy.
 
 - `src/components/`
   Reusable UI building blocks.
@@ -345,6 +324,7 @@ scene-specific workflows.
   - `layout.js`
     `src/internal/ui/` is the only shared home for page/store/handler
     orchestration that does not belong inside one page folder.
+    Do not move page-private support code there.
 
 - `src/deps/services/shared/`
   Shared handler-facing and internal service logic.
