+++ src/components/layoutEditorPreview/support/layoutEditorPreviewVariables.js
@@ -1,9 +1,9 @@
-import { getSystemVariableItems } from "../../../systemVariables.js";
+import { getSystemVariableItems } from "../../../internal/systemVariables.js";
 import {
   getRuntimeLayoutConditionItems,
   splitLayoutConditionFromWhen,
-} from "../../../layoutConditions.js";
-import { visitLayoutItemsWithFragments } from "./fragments.js";
+} from "../../../internal/layoutConditions.js";
+import { visitLayoutItemsWithFragments } from "./layoutEditorPreviewFragments.js";
 
 const PREVIEW_VARIABLE_TYPES = new Set(["boolean", "number", "string"]);
 const NORMAL_LIKE_LAYOUT_TYPES = new Set([
