+++ src/pages/layoutEditor/layoutEditor.store.js
@@ -9,7 +9,7 @@ import {
   selectLayoutEditorSelectedItem,
   toLayoutEditorContextMenuItems,
   toLayoutEditorExplorerItems,
-} from "../../internal/ui/layoutEditor/layoutEditorViewData.js";
+} from "./support/layoutEditorViewData.js";
 
 export const createInitialState = () => {
   return {
